python/requirements.txt
pyserial
pyautogui
time